requirements.txt
black==23.3.0
ruff==0.3.3
pytest==7.2.0
python-dotenv==1.0.1
numpy==1.26.4
matplotlib==3.8.3
scipy==1.12.0